refactor: 에러 메시지 디렉토리 이동, enum name 변경
